backend/requirements.txt
django==3.*
django-filter==2.*
djangorestframework==3.*
djangorestframework-api-key==2.*
djangorestframework_simplejwt==4.*
django-unixtimestampfield==0.4.0
get-docker-secret==1.*
gunicorn==20.*
markdown==3.*
psycopg2==2.*
ldap3==2.*
elastic-apm